core: rename core-varlink -> varlink
To make things consistent with dbus.[ch]
