+++ src/core/varlink.h
